refactor component
