+++ frontend/src/scenes/debug/DebugScene.tsx
@@ -3,76 +3,12 @@ import { PageHeader } from 'lib/components/PageHeader'
 import { LemonButton } from 'lib/lemon-ui/LemonButton'
 import { LemonLabel } from 'lib/lemon-ui/LemonLabel/LemonLabel'
 import { LemonSelect } from 'lib/lemon-ui/LemonSelect'
-import { HogQLDebug } from 'scenes/debug/HogQLDebug'
-import { Modifiers } from 'scenes/debug/Modifiers'
-import { QueryTabs } from 'scenes/debug/QueryTabs'
+import { DebugSceneQuery } from 'scenes/debug/DebugSceneQuery'
 import { SceneExport } from 'scenes/sceneTypes'
 
 import { stringifiedExamples } from '~/queries/examples'
-import { dataNodeLogic, DataNodeLogicProps } from '~/queries/nodes/DataNode/dataNodeLogic'
-import { QueryEditor } from '~/queries/QueryEditor/QueryEditor'
-import { DataNode, HogQLQuery, Node } from '~/queries/schema'
 
 import { debugSceneLogic } from './debugSceneLogic'
-
-interface QueryDebugProps {
-    queryKey: string
-    query: string
-    setQuery: (query: string) => void
-}
-function QueryDebug({ query, setQuery, queryKey }: QueryDebugProps): JSX.Element {
-    let parsed: Record<string, any> | null = null
-    try {
-        parsed = JSON.parse(query)
-    } catch (e) {
-        // do nothing
-    }
-
-    const dataNodeLogicProps: DataNodeLogicProps = {
-        query: (parsed ?? {}) as DataNode,
-        key: queryKey,
-        dataNodeCollectionId: queryKey,
-    }
-    const { response } = useValues(dataNodeLogic(dataNodeLogicProps))
-
-    return (
-        <>
-            {parsed && parsed?.kind === 'HogQLQuery' ? (
-                <HogQLDebug
-                    queryKey={queryKey}
-                    query={parsed as HogQLQuery}
-                    setQuery={(query) => setQuery(JSON.stringify(query, null, 2))}
-                />
-            ) : (
-                <div className="space-y-4">
-                    <QueryEditor
-                        query={query}
-                        setQuery={setQuery}
-                        aboveButton={
-                            <Modifiers
-                                setQuery={
-                                    parsed?.source
-                                        ? (query) => setQuery(JSON.stringify({ ...parsed, source: query }, null, 2))
-                                        : (query) => setQuery(JSON.stringify(query, null, 2))
-                                }
-                                query={parsed?.source ?? parsed}
-                                response={response}
-                            />
-                        }
-                    />
-                    {parsed ? (
-                        <QueryTabs
-                            query={parsed as Node}
-                            queryKey={queryKey}
-                            response={response}
-                            setQuery={(query) => setQuery(JSON.stringify(query, null, 2))}
-                        />
-                    ) : null}
-                </div>
-            )}
-        </>
-    )
-}
 
 export function DebugScene(): JSX.Element {
     const { query1, query2 } = useValues(debugSceneLogic)
@@ -124,11 +60,11 @@ export function DebugScene(): JSX.Element {
             />
             <div className="flex gap-2">
                 <div className="flex-1 w-1/2">
-                    <QueryDebug query={query1} setQuery={setQuery1} queryKey="hogql-debug-1" />
+                    <DebugSceneQuery query={query1} setQuery={setQuery1} queryKey="hogql-debug-1" />
                 </div>
                 {query2 ? (
                     <div className="flex-1 w-1/2">
-                        <QueryDebug query={query2} setQuery={setQuery2} queryKey="hogql-debug-2" />
+                        <DebugSceneQuery query={query2} setQuery={setQuery2} queryKey="hogql-debug-2" />
                     </div>
                 ) : null}
             </div>
